moved test runner features files to their own folder
